student-mentor-backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
google-adk==0.3.1
google-cloud-firestore==2.16.0
python-dotenv==1.0.1
firebase-admin==6.5.0
